# perf(stdlib): optimize pivot for common scenarios