--- a/internal/arrowutil/types.tmpldata
+++ b/internal/arrowutil/types.tmpldata
@@ -1,0 +1,47 @@
+[
+  {
+    "Name": "Int64",
+    "Type": "array.Int64",
+    "PrimitiveType": "int64",
+    "IsNumeric": true,
+    "IsComparable": true,
+    "Value": "Value",
+    "Append": "Append"
+  },
+  {
+    "Name": "Uint64",
+    "Type": "array.Uint64",
+    "PrimitiveType": "uint64",
+    "IsNumeric": true,
+    "IsComparable": true,
+    "Value": "Value",
+    "Append": "Append"
+  },
+  {
+    "Name": "Float64",
+    "Type": "array.Float64",
+    "PrimitiveType": "float64",
+    "IsNumeric": true,
+    "IsComparable": true,
+    "Value": "Value",
+    "Append": "Append"
+  },
+  {
+    "Name": "Boolean",
+    "Type": "array.Boolean",
+    "PrimitiveType": "bool",
+    "IsNumeric": false,
+    "IsComparable": false,
+    "Value": "Value",
+    "Append": "Append"
+  },
+  {
+    "Name": "String",
+    "Type": "array.Binary",
+    "PrimitiveType": "string",
+    "IsNumeric": false,
+    "IsComparable": true,
+    "Value": "ValueString",
+    "Append": "AppendString"
+  }
+]