… go.sum
… internal/arrowutil/builder.gen.go
… internal/arrowutil/iterator.gen.go
… internal/arrowutil/iterator.gen_test.go
… internal/execute/tableutil/buffered_test.go
… internal/tools/go.sum
… internal/types.tmpldata
… stdlib/universe/pivot.gen.go
… stdlib/universe/pivot.go
… stdlib/universe/pivot_test.go