@jsternberg 2019-11-14
Benchmark for this:

```
jsternberg:[~/g/s/g/i/flux] $ (perf/optimize-pivot *$) benchcmp old.txt new.txt
benchmark                  old ns/op       new ns/op      delta
BenchmarkPivot/1000-12     33557264224     5613306182     -83.27%

benchmark                  old allocs     new allocs     delta
BenchmarkPivot/1000-12     136673348      3813245        -97.21%

benchmark                  old bytes      new bytes      delta
BenchmarkPivot/1000-12     9367265416     3319302120     -64.56%
```

It likely still needs some tests, but this PR doesn't enable the code. I think we're likely going to need #2131 for this to work effectively so we can confidently enable this code in specific scenarios.