@jsternberg 2019-11-19
I updated this so it works with the new way that tools are versioned. This is ready for review again.