@jsternberg 2019-11-19
Test cases from the v1 version of pivot were adapted for v2. The v2 one only works if the column key is part of the group key and that the row key has no nulls so those tests were removed.