@jsternberg 2019-11-26
I think I probably do need to create tests for the new functionality that isn't being tested. I'll try to write those today. If this begins to get too delayed because I'm having difficulty writing the tests, we can split the work.