@jsternberg 2019-11-26
Wrote tests for arrowutil and the other stuff that's in here. I also cleaned up the iterators a bit. Previously, I had next be called after reading the value, but that's not really expected in typical Go APIs. I updated it so next is always called before reading the value. That made the merge code a bit different (and the previous api made it easier), but I don't think there's much of a meaningful difference and having the iterators with a weird API specifically for one algorithm seemed a bit strange to me.