@jsternberg 2019-11-26
I tried to add more patch coverage. I think part of the issue is a lot of the generated file is not actually tested because it generates the code paths for the row key to be a non-time column. I think I am just going to have to go with what I currently have for now. I believe the test coverage to be sufficient enough.